Lots of teleop fixes, among other things
I updated the registry to reflect what we want. Also, I updated the
Library and the Teleop to be more functional for the game. It now has
support for the dongler system, the motor for the hanging system (still
haven't done the servos, will do tomorrow), and now is driven by a
single stick, though we can easily switch back to independent movement
if wanted.